settings/requirements.txt
python-dotenv
PyYaml
jinja2
python-magic
# not sure about this one
pathlib

